server/requirements.txt
beautifulsoup4==4.9.3
breadability==0.1.20
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
docopt==0.6.2
emoji==1.4.2
Flask==2.0.1
Flask-Cors==3.0.10
gensim==4.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
mecab-python3==1.0.4
nltk==3.6.5
numpy==1.21.2
pandas==1.3.2
pycountry==20.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-Levenshtein==0.12.2
pytz==2021.1
regex==2021.10.8
requests==2.26.0
scipy==1.7.1
six==1.16.0
smart-open==5.2.0
soupsieve==2.2.1
sumy==0.8.1
tinysegmenter==0.4
tqdm==4.62.3
tzlocal==2.1
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==2.0.1
